+++ campus/src/app/app-routing.module.ts
@@ -56,8 +56,8 @@ const routes: Routes = [
     path: 'coursemy',
     loadChildren: () => import('./page/studyroompage/course-my/course-my.module').then(m => m.CourseMyPageModule)
   },  {
-    path: 'courselistadmin',
-    loadChildren: () => import('./page/studyroompage/course-list-admin/course-list-admin.module').then(m => m.CourseListAdminPageModule)
+    path: 'lessonlistadmin',
+    loadChildren: () => import('./page/studyroompage/lesson-list-admin/lesson-list-admin.module').then(m => m.LessonListAdminPageModule)
   },
   {
     path: 'projectsearch',
