+++ campus/src/app/page/studyroompage/lesson-list-admin/lesson-list-admin.page.scss
@@ -174,7 +174,7 @@ ion-col{
 .pagination {
   display: flex;
   justify-content: center; // 수평 가운데 정렬
-  margin-top: 20px; // 버튼과 위 요소 간의 간격 조정
+  margin-top: 5px; // 버튼과 위 요소 간의 간격 조정
 }
 
 .pagination-buttons {
